Implemented new version of ExampleBasedMatcher (and broke old version). The whole matching process was logically simplified, yet it provides more opportunities for matching with so-called "command template language". Added essential structure for command configurations used by ExampleBasedMatcher. Split new code for ExampleBasedMatcher into different files. Added a few functions to work with dictionaries, added new TemplateFiller class. Moved old code for ExampleBasedMatcher into different directory.
The only TODO is to (again) implement configurations loader for matcher.
