+++ src/ansible_matcher/example_based/commands/mkdir.json
