+++ src/ansible_matcher/example_based/commands/rm.json
